requirements.txt
django==4.2.29
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pytest==8.4.2
pytest-cov==7.1.0
pytest-django==4.11.1
